Fix canvas glitches
Fixed the issues of black edges on both sides of the dashboard canvas and screen flickering and tearing.

In some operating systems, there are issues with the SDL2 vsync/renderer driver. Their SDL2 is a newer version, and the internal renderer/backend implementation of SDL2 has changed. ICSim does not enforce SDL_RENDERER_PRESENTVSYNC, so in certain SDL2 versions, vertical synchronization is not enabled → resulting in rendering tearing.